Post Processor for Dialog 11 Control via Fusion 360
Template file is from Fusion 360 Library for Dialog 11. I made manual edits to the code for some of the drilling and tapping cycles. Not all drilling/reaming/boring cycles have been defined but drilling, deep drilling, tapping were working on my Deckel FP50 machine. DSI helped me make some edits to the cycles so that the lines of code for the cycles are inserted directly into the program each time they're used instead of referencing a sub routine. I did this because I could not figure out how to get my machine to import the sub programs with the main program so I would have to manually define them on the control. Ask if you have any questions.